// frontend/src/app/generated/sql_menu_page.ts
export interface MenuPage {
    /**
     * @generated from protobuf field: int32 menu_card_id = 1
     */
    menuCardId: number;
    /**
     * @generated from protobuf field: int32 menu_page_id = 2
     */
    menuPageId: number;
    /**
     * @generated from protobuf field: string local_name = 3
     */
    localName: string;
    /**
     * @generated from protobuf field: string chinese_name = 4
     */
    chineseName: string;
    /**
     * @generated from protobuf field: int32 page_button_size = 5
     */
    pageButtonSize: number;
    /**
     * @generated from protobuf field: string picture = 6
     */
    picture: string;
    /**
     * @generated from protobuf field: bool is_vertical_orientation = 7
     */
    isVerticalOrientation: boolean;
}
// ...
class MenuPage$Type extends MessageType<MenuPage> {
    constructor() {
        super("mp.MenuPage", [
            { no: 1, name: "menu_card_id", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 2, name: "menu_page_id", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 3, name: "local_name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 4, name: "chinese_name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 5, name: "page_button_size", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 6, name: "picture", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 7, name: "is_vertical_orientation", kind: "scalar", T: 8 /*ScalarType.BOOL*/ }
        ]);
    }
    create(value?: PartialMessage<MenuPage>): MenuPage {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.menuCardId = 0;
        message.menuPageId = 0;
        message.localName = "";
        message.chineseName = "";
        message.pageButtonSize = 0;
        message.picture = "";
        message.isVerticalOrientation = false;
        if (value !== undefined)
            reflectionMergePartial<MenuPage>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: MenuPage): MenuPage {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* int32 menu_card_id */ 1:
                    message.menuCardId = reader.int32();
                    break;
                case /* int32 menu_page_id */ 2:
                    message.menuPageId = reader.int32();
                    break;
                case /* string local_name */ 3:
                    message.localName = reader.string();
                    break;
                case /* string chinese_name */ 4:
                    message.chineseName = reader.string();
                    break;
                case /* int32 page_button_size */ 5:
                    message.pageButtonSize = reader.int32();
                    break;
                case /* string picture */ 6:
                    message.picture = reader.string();
                    break;
                case /* bool is_vertical_orientation */ 7:
                    message.isVerticalOrientation = reader.bool();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: MenuPage, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* int32 menu_card_id = 1; */
        if (message.menuCardId !== 0)
            writer.tag(1, WireType.Varint).int32(message.menuCardId);
        /* int32 menu_page_id = 2; */
        if (message.menuPageId !== 0)
            writer.tag(2, WireType.Varint).int32(message.menuPageId);
        /* string local_name = 3; */
        if (message.localName !== "")
            writer.tag(3, WireType.LengthDelimited).string(message.localName);
        /* string chinese_name = 4; */
        if (message.chineseName !== "")
            writer.tag(4, WireType.LengthDelimited).string(message.chineseName);
        /* int32 page_button_size = 5; */
        if (message.pageButtonSize !== 0)
            writer.tag(5, WireType.Varint).int32(message.pageButtonSize);
        /* string picture = 6; */
        if (message.picture !== "")
            writer.tag(6, WireType.LengthDelimited).string(message.picture);
        /* bool is_vertical_orientation = 7; */
        if (message.isVerticalOrientation !== false)
            writer.tag(7, WireType.Varint).bool(message.isVerticalOrientation);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message mp.MenuPage
 */
export const MenuPage = new MenuPage$Type();